+++ src/Common/Internal/DisposeAction.cs
