+++ src/Microsoft.Cci.Extensions/Microsoft.Cci.Extensions.csproj
@@ -1,4 +1,4 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
+<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <DisableImplicitAssetTargetFallback>true</DisableImplicitAssetTargetFallback>
     <TargetFrameworks>netstandard2.0</TargetFrameworks>
@@ -12,6 +12,10 @@
     <NoWarn>NU1701</NoWarn>
     <DefineConstants>$(DefineConstants);COREFX</DefineConstants>
   </PropertyGroup>
+
+  <ItemGroup>
+    <Compile Include="..\Common\Internal\DisposeAction.cs" Link="Internal\DisposeAction.cs"/>
+  </ItemGroup>
 
   <ItemGroup>
     <PackageReference Include="Microsoft.Cci" Version="4.0.0-rc3-24214-00" />
