+++ src/Microsoft.DotNet.AsmDiff/DiffDocument.cs
@@ -2,7 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Linq;
@@ -36,72 +35,5 @@ namespace Microsoft.DotNet.AsmDiff
                 return !hasLeft || hasRight;
             }
         }
-    }
-
-    public sealed class DiffLine
-    {
-        public DiffLineKind Kind { get; private set; }
-        public ReadOnlyCollection<DiffToken> Tokens { get; private set; }
-
-        public DiffLine(DiffLineKind kind, IEnumerable<DiffToken> tokens)
-        {
-            Kind = kind;
-            Tokens = new ReadOnlyCollection<DiffToken>(tokens.ToArray());
-        }
-
-        public override string ToString()
-        {
-            return string.Concat(Tokens);
-        }
-    }
-
-    public sealed class DiffToken
-    {
-        public DiffStyle Style { get; private set; }
-        public DiffTokenKind Kind { get; set; }
-        public string Text { get; private set; }
-
-        public DiffToken(DiffStyle style, DiffTokenKind kind, string text)
-        {
-            Style = style;
-            Text = text;
-            Kind = kind;
-        }
-
-        public override string ToString()
-        {
-            return Text;
-        }
-    }
-
-    public enum DiffLineKind
-    {
-        Same,
-        Added,
-        Removed,
-        Changed
-    }
-
-    [Flags]
-    public enum DiffStyle
-    {
-        None = 0x00,
-        Added = 0x01,
-        Removed = 0x02,
-        InterfaceMember = 0x04,
-        InheritedMember = 0x08,
-        NotCompatible = 0x10,
-    }
-
-    public enum DiffTokenKind
-    {
-        Text,
-        Symbol,
-        Identifier,
-        Keyword,
-        TypeName,
-        LineBreak,
-        Indent,
-        Whitespace
-    }
+    }    
 }
