+++ src/Microsoft.DotNet.AsmDiff/DiffLineKind.cs
@@ -1,0 +1,14 @@
+// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+namespace Microsoft.DotNet.AsmDiff
+{
+    public enum DiffLineKind
+    {
+        Same,
+        Added,
+        Removed,
+        Changed
+    }
+}
